refactor: 피드백 반영 #2
RunRace 테스트
메서드명 변경 isSize -> checkSize
패키지 변경
