+++ src/test/java/domain/RunRaceTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import service.RunRace;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
